Rename the example data file
