+++ apps/cookbook/src/app/examples/data-table-example/example-data.ts
